Task 2/requirements.txt
kafka-python
pandas
numpy
scikit-learn
pyflink